Add requests/limits to stack diagnostics Pod (#74)
Sets limits/requests to allow deployment of stack diagnostics Pods into namespaces that have quotas configured. The initContainer is set to run with memory/cpu 20Mi/100m and limit 2Gi/1 which matches the default JVM options the support diagnostics tool uses. The main container which does not work other than waiting for the extraction of data via tar is running with 16Mi/50m. A lower limit would be technically possible from the containers point of view but different container runtimes do not allow to go below a threshold value which can be as high as 6Mi for Docker for example.