Add support for elastic-agent diagnostics (#97)
Adds an option to run the elastic-agent diagnostics subcommand. This is disabled by default as it will include credentials in the archive, which is a limitation of the diagnostics command.

The main use case I have in mind, other than user installations, is the ECK operator e2e test jobs where I found it helpful to have the diagnostics around for analysing Elastic Agent related test failures. But it is quite tedious to extract the diagnostics manually in a containerised environment. With this functionality in place the diagnostics archives we have in CI will already contain Agent diagnostics.

For use in large scale Elastic Agent deployments we will probably have to make additional adjustments. The current implementation is synchronous handling one Agent after another which is going to take a long time when users have 100s or 1000s of Agents deployed.

Given that we have similar limitations with the support diagnostics we extract from Elasticsearch and Kibana I thought it might be an acceptable tradeoff for now. And we have #81 to address usage in large scale environments.