Run support-diagnostics as non-root user (#123)
This commit changes the job used to run the support-diagnostics 
to use a security context with a non-root user. The command 
arguments and the readiness probe are updated so that everything
is done under the output directory where the user has write 
permissions (the utility writes a startup.log file and we write
an empty ready file).